+++ src/pages/CategoryPage.jsx
@@ -1,0 +1,5 @@
+const CategoryPage = () => {
+  return <h1>카테고리 페이지</h1>;
+};
+
+export default CategoryPage;
