--- src/pages/CategoryPage/CategoryPage.jsx
@@ -1,7 +1,0 @@
-import * as S from '@/pages/CategoryPage/CategoryPage.style';
-
-const CategoryPage = () => {
-  return <S.Container>메인 페이지</S.Container>;
-};
-
-export default CategoryPage;
