+++ src/pages/MainPage.jsx
@@ -1,0 +1,5 @@
+const MainPage = () => {
+  return <h1>메인 페이지</h1>;
+};
+
+export default MainPage;
