--- src/pages/MainPage/MainPage.jsx
@@ -1,7 +1,0 @@
-import * as S from '@/pages/MainPage/MainPage.style';
-
-const MainPage = () => {
-  return <S.Container>메인 페이지</S.Container>;
-};
-
-export default MainPage;
