--- src/pages/MainPage/MainPage.style.js
@@ -1,3 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div``;
