+++ src/pages/MyPage.jsx
@@ -1,0 +1,5 @@
+const MyPage = () => {
+  return <h1>마이 페이지</h1>;
+};
+
+export default MyPage;
