--- src/pages/MyPage/MyPage.jsx
@@ -1,7 +1,0 @@
-import * as S from '@/pages/Mypage/MyPage.style';
-
-const MyPage = () => {
-  return <S.Container>메인 페이지</S.Container>;
-};
-
-export default MyPage;
