+++ src/pages/RecordPage.jsx
@@ -1,0 +1,5 @@
+const RecordPage = () => {
+  return <h1>기록 페이지</h1>;
+};
+
+export default RecordPage;
