--- src/pages/RecordPage/RecordPage.jsx
@@ -1,7 +1,0 @@
-import * as S from '@/pages/RecordPage/RecordPage.style';
-
-const RecordPage = () => {
-  return <S.Container>메인 페이지</S.Container>;
-};
-
-export default RecordPage;
