+++ src/pages/SplashPage.jsx
@@ -1,0 +1,5 @@
+const SplashPage = () => {
+  return <h1>스플래쉬 페이지</h1>;
+};
+
+export default SplashPage;
