--- src/pages/SplashPage/SplashPage.jsx
@@ -1,7 +1,0 @@
-import * as S from '@/pages/SplashPage/SplashPage.style';
-
-const SplashPage = () => {
-  return <S.Container>메인 페이지</S.Container>;
-};
-
-export default SplashPage;
