--- src/pages/TodoPage/TodoPage.jsx
@@ -1,7 +1,0 @@
-import * as S from '@/pages/TodoPage/TodoPage.style';
-
-const TodoPage = () => {
-  return <S.Container>투두 페이지</S.Container>;
-};
-
-export default TodoPage;
